Moving locking one step deeper
